+++ packages/connector/src/connectors/ethereum/common/web3connection.ts
@@ -1,6 +1,6 @@
 import type { Observable } from "rxjs"
 import { combineLatest } from "rxjs"
-import { map } from "rxjs/operators"
+import { distinctUntilChanged, map } from "rxjs/operators"
 import type { EthereumProviderConnectionResult } from "../domain"
 import type { ConnectionState } from "../../../connection-state"
 import { getStateConnected, getStateDisconnected } from "../../../connection-state"
@@ -19,6 +19,9 @@ export function connectToWeb3(
 		getChainId(provider),
 		getConnectedStatus(provider),
 	]).pipe(
+		distinctUntilChanged((c1, c2) => {
+			return c1[0] === c2[0] && c1[1] === c2[1] && c1[2] === c2[2]
+		}),
 		map(([address, chainId, status]) => {
 			if (status === "connected" && address) {
 				const wallet: EthereumProviderConnectionResult = {